split Evaluations tabs page into individual pages
